+++ js/server.js
@@ -1,10 +1,13 @@
 import { isEscape } from './utils.js';
+import { showSuccessMessage, showErrorMessage } from './messages.js';
 
 const API_URL = 'https://28.javascript.htmlacademy.pro/kekstagram';
-const SUCCESS_TEMPLATE = document.querySelector('#success');
-const ERROR_TEMPLATE = document.querySelector('#error');
-const body = document.body;
 
+/**
+ * Отправка данных на сервер
+ * @param {FormData} formData - данные формы
+ * @returns {Promise<boolean>} - результат отправки
+ */
 export async function sendDataToServer(formData) {
   try {
     const response = await fetch(API_URL, {
@@ -27,63 +30,10 @@ export async function sendDataToServer(formData) {
   }
 }
 
-function showSuccessMessage() {
-  const successElement = SUCCESS_TEMPLATE.content.querySelector('.success').cloneNode(true);
-  body.appendChild(successElement);
-
-  const messageEl = successElement;
-
-  const onClick = (e) => {
-    if (e.target.classList.contains('success__button')) {
-      removeMessage();
-    }
-  };
-
-  const onKeyDown = (e) => {
-    if (isEscape(e)) {
-      removeMessage();
-    }
-  };
-
-  function removeMessage() {
-    messageEl.remove();
-    document.removeEventListener('click', onClick);
-    document.removeEventListener('keydown', onKeyDown);
-  }
-
-  document.addEventListener('click', onClick);
-  document.addEventListener('keydown', onKeyDown);
-}
-
-function showErrorMessage() {
-
-  const errorElement = ERROR_TEMPLATE.content.querySelector('.error').cloneNode(true);
-  body.appendChild(errorElement);
-
-  const messageEl = errorElement;
-
-  const onClick = (e) => {
-    if (e.target.classList.contains('error__button')) {
-      removeMessage();
-    }
-  };
-
-  const onKeyDown = (e) => {
-    if (isEscape(e)) {
-      removeMessage();
-    }
-  };
-
-  function removeMessage() {
-    messageEl.remove();
-    document.removeEventListener('click', onClick);
-    document.removeEventListener('keydown', onKeyDown);
-  }
-
-  document.addEventListener('click', onClick);
-  document.addEventListener('keydown', onKeyDown);
-}
-
+/**
+ * Блокировка кнопки отправки формы
+ * @param {HTMLElement} button - кнопка формы
+ */
 export function blockSubmitButton(button) {
   if (button) {
     button.disabled = true;
@@ -91,6 +41,10 @@ export function blockSubmitButton(button) {
   }
 }
 
+/**
+ * Разблокировка кнопки отправки формы
+ * @param {HTMLElement} button - кнопка формы
+ */
 export function unblockSubmitButton(button) {
   if (button) {
     button.disabled = false;
@@ -98,6 +52,11 @@ export function unblockSubmitButton(button) {
   }
 }
 
+/**
+ * Создает объект FormData из формы
+ * @param {HTMLFormElement} form - форма
+ * @returns {FormData}
+ */
 export function createFormData(form) {
   return new FormData(form);
 }
